quise templatizar la funcion q lee los CSV pero no me salio, asi q la pase a float q es lo que usamos en el perceptron. Ademas agregue un file.close() en esa funcion, solo un detalle

+++ Perceptron/include/utils.h
@@ -9,7 +9,7 @@
 #include <cstdlib>
 #include <cassert>
 
-void parseCSV(std::string filename, std::vector<std::vector<double> > & X);
+void parseCSV(std::string filename, std::vector<std::vector<float> > & X);
 
 // Debo definir las funciones con templates en los headers porque sino no linkea
